requirements.txt
# Dev requirements
isort==5.5.3
black==20.8b1
pre-commit
bumpversion==0.6.0
Sphinx==3.1.1
sphinxcontrib-httpdomain==1.7.0
sphinx-rtd-theme==0.4.3

# Test requirements
codecov==2.1.9
coverage==5.3
tox==3.15.2
pytest==5.3.5
pytest-runner>=2.11.1
pytest-django==3.8.0
pytest-kind==20.5.3

# Runtime
pip==20.1.1
watchdog==0.10.3


docker==4.3.1
Django==2.2.13
django-livesync==0.4
django-constance==2.6.0
django-picklefield==1.0.0
djangorestframework==3.9.1
kubernetes==11.0.0
rq-scheduler==0.8.3
psycopg2==2.7.5
pid==2.2.0
-e git+https://github.com/istrategylabs/django-rq-scheduler.git@fa76c6ac71068c304aa14ab9df7b13a4b972132e#egg=django-rq-scheduler

